get time printed
