+++ planet/src/main/java/com/google/planet/data/ItineraryGenerator.java
@@ -19,6 +19,19 @@ import java.util.Arrays;
 import java.util.ArrayList;
 import java.util.Collections;
 
+
+import com.google.maps.GeoApiContext;
+import com.google.maps.DistanceMatrixApi;
+import com.google.maps.GaeRequestHandler;
+import com.google.gson.GsonBuilder;
+import com.google.maps.model.GeocodingResult;
+import com.google.maps.model.DistanceMatrix;
+import com.google.maps.model.DistanceMatrixRow;
+import com.google.maps.model.DistanceMatrixElement;
+import com.google.maps.model.Duration;
+import com.google.maps.DistanceMatrixApiRequest;
+import com.google.maps.*;
+
 public final class ItineraryGenerator {
     public List<ItineraryItem> generateItinerary(List<Event> events) { 
         // Return an empty list if events are empty or only contains the starting point event
@@ -33,6 +46,7 @@ public final class ItineraryGenerator {
 
         Collections.sort(events, Event.SortByOrder);
         int [][] travelTimeGraph = getTravelTimeGraph(events);
+        getEventsDistanceMatrix();
         List<ItineraryItem> items = scheduleItineraryInOrder(events, START, END, travelTimeGraph);
         return items;
     }
@@ -40,7 +54,30 @@ public final class ItineraryGenerator {
     // Function that creates a graph with the events as vertices and the travelling
     // time between the events as edges.
     // For the prototype, all travelling times are set to 15 minutes 
-    public int[][] getTravelTimeGraph(List<Event> events) {
+    private boolean getEventsDistanceMatrix() {
+        String GoogleApiKey = "AIzaSyDK36gDoYgOj4AlbCqh1IuaUuTlcpKF0ns";
+        String[] origins = new String[] {"Vancouver+BC", "San+Francisco" };
+        String[] destinations = new String[] {"Vancouver+BC", "San+Francisco"};
+
+        GeoApiContext context = new GeoApiContext.Builder(new GaeRequestHandler.Builder())
+                            .apiKey(GoogleApiKey)
+                            .build();
+       try {
+            DistanceMatrix result = DistanceMatrixApi.getDistanceMatrix(context,
+                                                         origins,
+                                                         destinations).await();
+            DistanceMatrixRow[] rows = result.rows;
+            DistanceMatrixElement[] elements = rows[0].elements;
+            Duration duration = elements[1].duration;
+            System.out.println(duration.humanReadable);
+            return true;
+        } catch (Exception e) {
+            System.out.println(e);
+            return false;
+        }
+    }
+
+    private int[][] getTravelTimeGraph(List<Event> events) {
         int numOfNodes = events.size();
         int[][] travelTimeGraph = new int[numOfNodes][numOfNodes];
         for (int i = 0; i < travelTimeGraph.length; i++) {
@@ -56,7 +93,7 @@ public final class ItineraryGenerator {
     }
 
     // Function that creates an itinerary by scheduling each event in order.
-    public List<ItineraryItem> scheduleItineraryInOrder(List<Event> events, int openningTime, int endingTime, 
+    private List<ItineraryItem> scheduleItineraryInOrder(List<Event> events, int openningTime, int endingTime, 
                                                         int[][] travelTimeGraph) {
         List<ItineraryItem> items = new ArrayList<>();
         int start = openningTime;
