+++ planet/src/main/webapp/js/authentication.js
@@ -25,6 +25,7 @@ const firebaseConfig = {
     measurementId: 'G-9W48MVBXLE'
 };
 let currentUser;
+let database;
 
 // Initialize the Firebase Application
 firebase.initializeApp(firebaseConfig);
